+++ Modules/Intent.Modules.AI.UnitTests/Tasks/GenerateUnitTestsWithAITask.cs
@@ -11,8 +11,7 @@ using Intent.IArchitect.Agent.Persistence.Output;
 using Intent.Metadata.Models;
 using Intent.Modelers.Domain.Api;
 using Intent.Modelers.Services.Api;
-using Intent.Modules.AI.ChatDrivenDomain.Settings;
-using Intent.Modules.AI.ChatDrivenDomain.Utils;
+using Intent.Modules.Common.AI;
 using Intent.Modules.VisualStudio.Projects.Api;
 using Intent.Plugins;
 using Intent.Registrations;
@@ -36,6 +35,7 @@ public class GenerateUnitTestsWithAITask : IModuleTask
     private readonly ISolutionConfig _solution;
     private readonly IOutputRegistry _outputRegistry;
     private IApplicationConfig _applicationConfig;
+    private readonly IntentSemanticKernelFactory _intentSemanticKernelFactory;
 
     private static readonly JsonSerializerOptions SerializerOptions = new()
     {
@@ -46,12 +46,14 @@ public class GenerateUnitTestsWithAITask : IModuleTask
         IApplicationConfigurationProvider applicationConfigurationProvider,
         IMetadataManager metadataManager,
         ISolutionConfig solution,
-        IOutputRegistry outputRegistry)
+        IOutputRegistry outputRegistry,
+            IUserSettingsProvider userSettingsProvider)
     {
         _applicationConfigurationProvider = applicationConfigurationProvider;
         _metadataManager = metadataManager;
         _solution = solution;
         _outputRegistry = outputRegistry;
+        _intentSemanticKernelFactory = new IntentSemanticKernelFactory(userSettingsProvider);
     }
 
     public string TaskTypeId => "Intent.Modules.AI.UnitTests.Generate";
@@ -66,7 +68,7 @@ public class GenerateUnitTestsWithAITask : IModuleTask
         _applicationConfig = _solution.GetApplicationConfig(applicationId);
 
         Logging.Log.Info($"Args: {string.Join(",", args)}");
-        var kernel = BuildSemanticKernel();
+        var kernel = _intentSemanticKernelFactory.BuildSemanticKernel();
         var chatResponseFormat = CreateJsonSchemaFormat();
         var executionSettings = new OpenAIPromptExecutionSettings
         {
@@ -170,131 +172,23 @@ public async Task Handle_Should_Create_Buyer_When_Command_Is_Valid()
         return prompt;
     }
 
-    private FileChange[] GetInputFiles(IElement element)
+    private List<ICodebaseFile> GetInputFiles(IElement element)
     {
-        var basePath = Path.GetFullPath(Path.Combine(_applicationConfig.DirectoryPath, _applicationConfig.OutputLocation));
-        var correlations = MetadataOutputFileCorrelationsPersistable.TryLoad(_applicationConfig.FilePath);
-        var outputLog = OutputLogPersistable.TryLoad(Path.Combine(_applicationConfig.DirectoryPath, $"{Path.GetFileNameWithoutExtension(_applicationConfig.FilePath)}.{OutputLogPersistable.FileExtension}"));
-
-        var fileMap = correlations!.Files
-            .SelectMany(x => x.Models, (file, model) => (File: file, Model: model))
-            .GroupBy(x => x.Model.Id)
-            .ToDictionary(x => x.Key, x => x.Select(i => Path.Combine(basePath, i.File.RelativePath)).ToList());
-
-        var correlatedFiles = CorrelatedFiles(fileMap, element.Id);
-
-
-        var files = new List<FileChange>();
-        var queryHandlerFile = correlatedFiles.First(x => x.FileName.EndsWith("Handler"));
-        files.Add(new FileChange(Path.GetRelativePath(basePath, queryHandlerFile.Path), "The Handler Class", queryHandlerFile.FileText));
-        var otherCorrelatedFiles = correlatedFiles.Where(x => !x.FileName.EndsWith("Handler") && Path.GetExtension(x.Path) == ".cs");
-        files.AddRange(otherCorrelatedFiles.Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "File supporting the Handler class", x.FileText)));
-        
+        var codebaseAccessor = _applicationConfigurationProvider.GetTrackedCodebaseFiles();
+        var inputFiles = codebaseAccessor.GetFilesForMetadata(element).ToList();
         if (element.TypeReference.ElementId != null)
         {
-            var returnTypeFile = CorrelatedFiles(fileMap, element.TypeReference.ElementId);
-            files.AddRange(returnTypeFile.Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "The return type for the Handler class", x.FileText)));
+            inputFiles.AddRange(codebaseAccessor.GetFilesForMetadata(element.TypeReference.Element));
         }
 
+        inputFiles.AddRange(codebaseAccessor.GetFilesForTemplate("Intent.EntityFrameworkCore.Repositories.EFRepositoryInterface"));
+        inputFiles.AddRange(codebaseAccessor.GetFilesForTemplate("Intent.EntityFrameworkCore.Repositories.RepositoryBase"));
+        inputFiles.AddRange(codebaseAccessor.GetFilesForTemplate("Intent.Entities.NotFoundException"));
+        inputFiles.AddRange(codebaseAccessor.GetFilesForTemplate("Intent.VisualStudio.Projects.VisualStudioSolution"));
 
-        foreach(var field in element.ChildElements) {
-            var returnTypeFile = CorrelatedFiles(fileMap, field.TypeReference.ElementId);
-            files.AddRange(returnTypeFile.Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "Data Transfer Objects for request object", x.FileText)));
-        }
-
-        files.AddRange(GetRelatedEntities(element)
-            .SelectMany(model => CorrelatedFiles(fileMap, model.Id)
-                .Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "Related entity persistence support file", x.FileText))));
-
-        if (outputLog != null)
-        {
-            outputLog.Load();
-            var outputLogFileMap = outputLog!.FileLogs
-                .GroupBy(x => x.CorrelationId)
-                .ToDictionary(x => x.Key, x => x.Select(i => i.FilePath).ToList());
-                ;
-            files.AddRange(CorrelatedFiles(outputLogFileMap, "Intent.EntityFrameworkCore.Repositories.EFRepositoryInterface")
-                .Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "EF Repository Interface", x.FileText)));
-            files.AddRange(CorrelatedFiles(outputLogFileMap, "Intent.EntityFrameworkCore.Repositories.RepositoryBase")
-                .Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "EF Repository Base Implementation", x.FileText)));
-            files.AddRange(CorrelatedFiles(outputLogFileMap, "Intent.Entities.NotFoundException")
-                .Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "NotFoundException", x.FileText)));
-            foreach (var package in _metadataManager.VisualStudio(_applicationConfig.Id).Packages)
-            {
-                files.AddRange(CorrelatedFiles(outputLogFileMap, $"Intent.VisualStudio.Projects.VisualStudioSolution#{package.Id}")
-                    .Select(x => new FileChange(Path.GetRelativePath(basePath, x.Path), "Visual Studio Solution File", x.FileText)));
-            }
-        }
-
-        return files.ToArray();
+        return inputFiles;
     }
-    /**
-## Adding code to the repository:
-* Ignore this unless you plant to use the repository method in the handler.
-* If there is no repository method that would be able to fetch out the required results in a performant way, then you may update the entity repository concrete and its interface with an appropriate method. You may need to introduce a Domain Object as the return type of these methods.
-* In this case, if you added any methods, then you must add the [IntentIgnore] attribute to that method.
-* Only add repository methods if one doesn't already exist that is appropriate and follows best practices.
-* The repository interfaces are in the Domain / Core layer and you so cannot reference types in the other layers. This is important for if you created Domain Objects as a return type for the new repository.
-
-     */
-
-    private Kernel BuildSemanticKernel()
-    {
-        var settings = _applicationConfigurationProvider.GetSettings().GetChatDrivenDomainSettings();
-        var model = string.IsNullOrWhiteSpace(settings.Model()) ? "gpt-4o" : settings.Model();
-        var apiKey = settings.APIKey();
-        // Create the Semantic Kernel instance with your LLM service.
-        // Replace <your-openai-key> with your actual OpenAI API key and adjust the model name as needed.
-        var builder = Kernel.CreateBuilder();
-        builder.Services.AddLogging(b => b.AddProvider(new SoftwareFactoryLoggingProvider()).SetMinimumLevel(LogLevel.Trace));
-
-        switch (settings.Provider().AsEnum())
-        {
-            case ChatDrivenDomainSettings.ProviderOptionsEnum.OpenAi:
-                if (string.IsNullOrWhiteSpace(apiKey))
-                {
-                    apiKey = Environment.GetEnvironmentVariable("OPENAI_API_KEY");
-                }
-
-                builder.Services.AddOpenAIChatCompletion(
-                    modelId: model,
-                    apiKey: apiKey ?? throw new Exception("No API Key defined. Locate the ChatDrivenDomainSettings App Settings or set the OPENAI_API_KEY environment variable."));
-                break;
-            case ChatDrivenDomainSettings.ProviderOptionsEnum.AzureOpenAi:
-                if (string.IsNullOrWhiteSpace(apiKey))
-                {
-                    apiKey = Environment.GetEnvironmentVariable("AZURE_OPENAI_API_KEY");
-                }
-
-                builder.Services.AddAzureOpenAIChatCompletion(
-                    deploymentName: settings.DeploymentName(),
-                    endpoint: settings.APIUrl(),
-                    apiKey: apiKey ?? throw new Exception("No API Key defined. Locate the ChatDrivenDomainSettings App Settings or set the AZURE_OPENAI_API_KEY environment variable."),
-                    modelId: model);
-                break;
-            case ChatDrivenDomainSettings.ProviderOptionsEnum.Ollama:
-#pragma warning disable SKEXP0070
-                builder.Services.AddOllamaChatCompletion(
-                    new OllamaApiClient(
-                        new HttpClient
-                        {
-                            Timeout = TimeSpan.FromMinutes(10),
-                            BaseAddress = new Uri(settings.APIUrl())
-                        },
-                        model)
-                );
-#pragma warning restore SKEXP0070
-                break;
-            default:
-                throw new ArgumentOutOfRangeException();
-        }
-
-        // Create a JSON schema format based on the FileChange array.
-        // This helps ensure that the LLM returns a valid JSON array conforming to our expected structure.
-        var kernel = builder.Build();
-        return kernel;
-    }
-
+   
     private static ChatResponseFormat CreateJsonSchemaFormat()
     {
         return ChatResponseFormat.CreateJsonSchemaFormat(jsonSchemaFormatName: "movie_result",
@@ -320,36 +214,6 @@ public async Task Handle_Should_Create_Buyer_When_Command_Is_Valid()
                                               }
                                               """),
             jsonSchemaIsStrict: true);
-    }
-
-    private static string Fail(string reason)
-    {
-        Logging.Log.Failure(reason);
-        var errorObject = new { errorMessage = reason };
-        var json = JsonSerializer.Serialize(errorObject, SerializerOptions);
-        return json;
-    }
-
-    private static List<(string Path, string FileName, string FileText)> CorrelatedFiles(Dictionary<string, List<string>> fileMap, string modelId)
-    {
-        if (fileMap.ContainsKey(modelId))
-        {
-            return fileMap[modelId].Select(x => (Path: x, FileName: Path.GetFileNameWithoutExtension(x), FileText: File.ReadAllText(x))).ToList();
-        }
-
-        return [];
-    }
-
-    private IEnumerable<ClassModel> GetRelatedEntities(IElement element)
-    {
-        var queriedEntity = element.AssociatedElements.FirstOrDefault(x => x.TypeReference.Element.IsClassModel())
-            ?.TypeReference.Element.AsClassModel();
-        if (queriedEntity == null)
-        {
-            return [];
-        }
-        var relatedClasses = new[] { queriedEntity }.Concat(queriedEntity.AssociatedClasses.Where(x => x.Class != null).Select(x => x.Class));
-        return relatedClasses;
     }
 
     public class FileChangesResult